SmolAgent/pyproject.toml
[project]
name = "smolagents-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.23.0",
    "smolagents[litellm,toolkit]>=1.19.0",
    "tavily-python>=0.7.9",
]
